chore: rename devicetree -> fdt
